update to use long double and long double complex

+++ include/mandelbrot_parameters.h
@@ -2,7 +2,7 @@
 #define MANDELBROT_PARAMETERS_H
 
 
-extern const double ESCAPE_RADIUS;
+extern const long double ESCAPE_RADIUS;
 
 
 #endif
